[ETCM-331] Extend Ip limit tracking for JsonRpc and not just the faucet
