+++ src/universal/conf/faucet.conf
@@ -95,6 +95,19 @@ mantis {
         # Domains allowed to query RPC endpoint. Use "*" to enable requests from
         # any domain.
         cors-allowed-origins = []
+
+        # Ip Limit tracking for JSON-RPC requests
+        # Limits the amount of request the same ip can perform in a given amount of time
+        ip-restriction {
+          # If enabled, restrictions are applied
+          enabled = false
+
+          # Time that should pass between requests
+          min-request-interval = 10.seconds
+
+          # Size of stored timestamps for requests made from each ip
+          latest-timestamp-cache-size = 1024
+        }
       }
 
       ipc {
